Added non-blocking mode to files including tests.
We now support a non-blocking mode for read/write calls with files. This
can be enabled via set_blocked().

To support that, the file protocol was extended to establish an
additional communication channel between server and client that is used
to notify the client of events. The client needs to explicitly request
these events so that the server knows when to notify the client and
when not to. Therefore, set_blocked() establishes this communication
channel. In non-blocking mode, the read/write methods don't ask the
server for the next input/output data, but instead request the
corresponding event from the server (if not already done) and return
WOULD_BLOCK to the caller. The caller can then come back later and ask
again. As soon as the event is received, the server is asked for the
next input/output data. This can, in rare circumstances, also return
WOULD_BLOCK, but typically the server will immediately deliver the
response.

So far, only vterm and pipes support the non-blocking mode. m3fs does
not support it yet, because for the memory backend there are no delayed
responses, so that the "blocking" mode is fine. We should however add
it later for the disk backend.

+++ src/apps/bench/accelchain/indirect.cc
@@ -70,10 +70,10 @@ void chain_indirect(Reference<File> in, Reference<File> out, size_t num, CycleDu
         acts[i]->start();
 
     size_t total = 0, seen = 0;
-    size_t count = in->read(buffer.get(), BUF_SIZE);
+    ssize_t count = in->read(buffer.get(), BUF_SIZE);
 
     // label 0 is special; use 1..n
-    accels[0]->write(buffer.get(), count);
+    accels[0]->write(buffer.get(), static_cast<size_t>(count));
     accels[0]->start(InDirAccel::Operation::COMPUTE, static_cast<size_t>(count), comptime, 1);
     ops[0] = InDirAccel::Operation::COMPUTE;
     total += static_cast<size_t>(count);
