+++ src/apps/bench/ycsb/lvldbserver/udp_handler.cc
@@ -53,9 +53,9 @@ UDPOpHandler::UDPOpHandler(NetworkManager &nm, const char *workload, m3::IpAddr 
     wl_size = 0;
     {
         FileRef wl_file(workload, FILE_R);
-        size_t len;
+        ssize_t len;
         while((len = wl_file->read(wl_buffer + wl_size, MAX_FILE_SIZE - wl_size)) > 0)
-            wl_size += len;
+            wl_size += static_cast<size_t>(len);
     }
 
     UNUSED uint64_t total_preins = static_cast<uint64_t>(wl_read4b());
